+++ test/rails_app/config/initializers/devise.rb
@@ -73,7 +73,7 @@ Devise.setup do |config|
   # ==> Configuration for :database_authenticatable
   # For bcrypt, this is the cost for hashing the password and defaults to 10. If
   # using other encryptors, it sets how many times you want the password re-encrypted.
-  config.stretches = Rails.env.test? ? 1 : 10
+  config.stretches = Rails.env.test? ? 4 : 10
 
   # ==> Configuration for :confirmable
   # The time you want to give your user to confirm their account. During this time